# api/openapi.yaml
openapi: 3.0.4
info:
  title: Favorite Game Gauntlet
  version: 0.9.0
paths:
...
  /games/wishlist:
    post:
      operationId: addWishlistGame
      tags: [ Games ]
      summary: Add game wishlist
      description: |
        Add the game you want to play here.<br>
        You can use another command to get a random game from this list.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WishlistGame'
      responses:
        '204':
          description: Wishlist game was saved successfully.
        '401':
          description: Authentication required.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    get:
      operationId: getWishlistGames
      tags: [ Games ]
      summary: Get game wishlist
      description: |
        You can see the games you might get here.<br>
        You can use another command to get a random game from this list.
      responses:
        '200':
          description: Wishlist games were retrieved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WishlistGames'
        '401':
          description: Authentication required.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /games/roll:
    post:
      operationId: makeGameRoll
      tags: [ Games ]
      summary: Select a random game from the wishlist
      description: |
        Select a random game from your unplayed list.<br>
        The selected game will become the current game.<br>
        You will then be able to start the timer.
      responses:
        '200':
          description: A game roll was performed and the selected game returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Game'
        '401':
          description: Authentication required.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: No unplayed games available for rolling.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Game roll not allowed due to current game state.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /games/current:
    get:
      operationId: getCurrentGame
      tags: [ Games ]
      summary: Get the current game
      description: |
        You can find out information and status of the current game.
      responses:
        '200':
          description: The user’s current game was retrieved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Game'
        '401':
          description: Authentication required.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Current game not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /games/current/finish:
    post:
      operationId: finishCurrentGame
      tags: [ Games ]
      summary: Complete the current game
      description: |
        You can complete the current game if you have completed it according to the FGG rules.<br>
        When you complete the game, you will receive a bonus that depends on the number of hours spent playing.
      responses:
        '204':
          description: Current game was marked as finished successfully.
        '401':
          description: Authentication required.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: No current game available to finish.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Game finishing not allowed due to lack of completed timers for the game.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /games/current/cancel:
    post:
      operationId: cancelCurrentGame
      tags: [ Games ]
      summary: Cancel the current game
      description: |
        You can cancel the current game if you do not intend to complete it for any reason.<br>
        If you cancel a game, you will receive a fixed penalty.
      responses:
        '204':
          description: Current game was marked as cancelled successfully.
        '401':
          description: Authentication required.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: No current game available to cancel.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /games/history:
    get:
      operationId: getGameHistory
      tags: [ Games ]
      summary: Get a list of all games played
      description: |
        You can get a complete list of all the games that were randomly selected from your wishlist,<br>
        as well as all the information about them.
      responses:
        '200':
          description: Game history was retrieved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Games'
        '401':
          description: Authentication required.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /points/territory-point-history:
    get:
      operationId: getTerritoryPointHistory
      tags: [ Points ]
      responses:
        '200':
          description: Success.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TerritoryPointChanges'
        '401':
          description: Authentication required.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /points/free-point-history:
    get:
      operationId: getFreePointHistory
      tags: [ Points ]
      responses:
        '200':
          description: Success.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FreePointChanges'
        '401':
          description: Authentication required.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /points/territory-points:
    post:
      operationId: changeTerritoryPoints
      tags: [ Points ]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                changeSource:
                  type: string
                changeValue:
                  type: integer
                  format: int
      responses:
        '200':
          description: Success.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TerritoryPointChange'
        '401':
          description: Authentication required.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: The points can't change on this value.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /points/free-points:
    post:
      operationId: changeFreePoints
      tags: [ Points ]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                changeSource:
                  type: string
                changeValue:
                  type: integer
                  format: int
      responses:
        '200':
          description: Success.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FreePointChange'
        '401':
          description: Authentication required.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: The points can't change on this value.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /points/territory-hours:
    post:
      operationId: changeTerritoryHours
      tags: [ Points ]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                changeValue:
                  type: integer
                  format: int
      responses:
        '200':
          description: Success.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Points'
        '401':
          description: Authentication required.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: The points can't change on this value.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /points/experience-points:
    post:
      operationId: changeExperiencePoints
      tags: [ Points ]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                changeValue:
                  type: integer
                  format: int
      responses:
        '200':
          description: Success.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Points'
        '401':
          description: Authentication required.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: The points can't change on this value.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /games/unplayed:
    post:
      operationId: addUnplayedGames
      tags: [ Games ]
      deprecated: true
      summary: Add game wishlist
      description: |
        Add the games you want to play here.<br>
        You can use another command to get a random game from this list.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WishlistGames'
      responses:
        '200':
          description: Unplayed games were saved successfully.
        '401':
          description: Authentication required.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    get:
      operationId: getUnplayedGames
      tags: [ Games ]
      deprecated: true
      summary: Get game wishlist
      description: |
        You can see the games you might get here.<br>
        You can use another command to get a random game from this list.
      responses:
        '200':
          description: Unplayed games were retrieved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WishlistGames'
        '401':
          description: Authentication required.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
...
  schemas:
...
    Name:
      type: string
      pattern: '^.+$'
      minLength: 1
      maxLength: 70
...
    Duration:
      type: string
      format: duration
      example: PT2h10m5s
      description: The duration notation as defined by ISO 8601
    Points:
      type: integer
      format: int
      minimum: 0
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
      required: [ code, message ]
...
    WishlistGame:
      type: object
      properties:
        name:
          $ref: '#/components/schemas/Name'
      required: [ name ]
    WishlistGames:
      type: array
      items:
        $ref: '#/components/schemas/WishlistGame'
    Game:
      type: object
      properties:
        name:
          $ref: '#/components/schemas/Name'
        state:
          type: string
          enum: [ 'started', 'finished', 'cancelled' ]
        timeSpent:
          $ref: '#/components/schemas/Duration'
        finishDate:
          type: string
          format: date-time
      required: [ name, state, timeSpent ]
    Games:
      type: array
      items:
        $ref: '#/components/schemas/Game'
...
    VictoryPointChange:
      type: object
      properties:
        changeSource:
          type: string
        changeValue:
          type: integer
          format: int
        actualChangeValue:
          type: integer
          format: int
        finalValue:
          $ref: '#/components/schemas/Points'
        changeDate:
          type: string
          format: date-time
      required: [ changeSource, changeValue, finalValue, changeDate ]
    TerritoryPointChange:
      allOf:
        - $ref: '#/components/schemas/VictoryPointChange'
        - type: object
          properties:
            changeSource:
              enum: [ 'territory-obtaining', 'territory-loss', 'other' ]
    TerritoryPointChanges:
      type: array
      items:
        $ref: '#/components/schemas/TerritoryPointChange'
    FreePointChange:
      allOf:
        - $ref: '#/components/schemas/VictoryPointChange'
        - type: object
          properties:
            changeSource:
              enum: [ 'own-wheel-effect', 'other-wheel-effect', 'quest', 'base-teleport', 'sandstorm', 'other' ]
            wheelEffectName:
              $ref: '#/components/schemas/Name'
    FreePointChanges:
      type: array
      items:
        $ref: '#/components/schemas/FreePointChange'